[GUI] New scripts structure
