+++ gui/model/benchmark_config/benchmark_config.py
@@ -1,151 +1,7 @@
 from xml.dom import minidom
-
-
-class Model:
-    def __init__(self, task=None, name=None, precision=None, framework=None, model_path=None, weights_path=None):
-        self.task = task
-        self.name = name
-        self.precision = precision
-        self.framework = framework
-        self.model_path = model_path
-        self.weights_path = weights_path
-
-    def __eq__(self, other):
-        return self.task == other.task and self.name == other.name and self.precision == other.precision and \
-               self.framework == other.framework and self.model_path == other.model_path and self.weights_path == other.weights_path
-
-    def get_str(self):
-        return self.task + ';' + self.name + ';' + self.precision + ';' + self.framework + ';' + self.model_path + ';' \
-               + self.weights_path
-
-
-class Dataset:
-    def __init__(self, name=None, path=None):
-        self.name = name
-        self.path = path
-
-    def __eq__(self, other):
-        return self.name == other.name and self.path == other.path
-
-    def get_str(self):
-        return self.name + ';' + self.path
-
-
-class Test:
-    def __init__(self, model=None, dataset=None, framework=None, batch_size=None, device=None, iter_count=None,
-                 test_time_limit=None, mode=None, extension=None, async_req_count=None, thread_count=None,
-                 stream_count=None, channel_swap=None, mean=None, input_scale=None):
-        self.model = model
-        self.dataset = dataset
-        self.framework = framework
-        self.batch_size = batch_size
-        self.device = device
-        self.iter_count = iter_count
-        self.test_time_limit = test_time_limit
-        self.mode = mode
-        self.extension = extension
-        self.async_req_count = async_req_count
-        self.thread_count = thread_count
-        self.stream_count = stream_count
-        self.channel_swap = channel_swap
-        self.mean = mean
-        self.input_scale = input_scale
-
-
-class RemoteComputer:
-    def __init__(self, ip=None, login=None, password=None, os=None, path_to_ftp_client=None, benchmark_config=None,
-                 log_file=None, res_file=None):
-        self.ip = ip
-        self.login = login
-        self.password = password
-        self.os = os
-        self.path_to_ftp_client = path_to_ftp_client
-        self.benchmark_config = benchmark_config
-        self.log_file = log_file
-        self.res_file = res_file
-
-
-class DeployComputer:
-    def __init__(self, ip=None, login=None, password=None, os=None, download_folder=None):
-        self.ip = ip
-        self.login = login
-        self.password = password
-        self.os = os
-        self.download_folder = download_folder
-
-
-class Models:
-    def __init__(self):
-        self.__models = []
-
-    def get_models(self):
-        return self.__models
-
-    def get_models_list(self):
-        models_list = []
-        for model in self.__models:
-            models_list.append(model.get_str())
-        return models_list
-
-    def set_models(self, models):
-        self.__models.clear()
-        for model in models:
-            if model not in self.__models:
-                self.__models.append(model)
-
-    def add_model(self, task, name, precision, framework, model_path, weights_path):
-        self.__models.append(Model(task, name, precision, framework, model_path, weights_path))
-
-    def change_model(self, row, task, name, precision, framework, model_path, weights_path):
-        self.__models[row] = Model(task, name, precision, framework, model_path, weights_path)
-
-    def delete_model(self, index):
-        self.__models.pop(index)
-
-    def delete_models(self, indexes):
-        for index in indexes:
-            if index < len(self.__models):
-                self.delete_model(index)
-
-    def clear(self):
-        self.__models.clear()
-
-
-class Data:
-    def __init__(self):
-        self.__data = []
-
-    def get_data(self):
-        return self.__data
-
-    def get_data_list(self):
-        data_list = []
-        for data in self.__data:
-            data_list.append(data.get_str())
-        return data_list
-
-    def set_data(self, data):
-        self.__data.clear()
-        for dataset in data:
-            if dataset not in self.__data:
-                self.__data.append(dataset)
-
-    def add_dataset(self, name, path):
-        self.__data.append(Dataset(name, path))
-
-    def change_dataset(self, row, name, path):
-        self.__data[row] = Dataset(name, path)
-
-    def delete_dataset(self, index):
-        self.__data.pop(index)
-
-    def delete_data(self, indexes):
-        for index in indexes:
-            if index < len(self.__data):
-                self.delete_dataset(index)
-
-    def clear(self):
-        self.__data.clear()
+from model.models.model import Model
+from model.data.dataset import Dataset
+from model.benchmark_config.test import Test
 
 
 class BenchmarkConfig:
@@ -433,186 +289,3 @@ class BenchmarkConfig:
             CONFIG_FRAMEWORK_INDEPENDENT_TAG.appendChild(CONFIG_MEAN_TAG)
             CONFIG_FRAMEWORK_INDEPENDENT_TAG.appendChild(CONFIG_INPUT_SCALE_TAG)
         return CONFIG_FRAMEWORK_INDEPENDENT_TAG
-
-class RemoteConfig:
-    def __init__(self):
-        self.__computers = []
-
-    def get_computers(self):
-        return self.__computers
-
-    def add_computer(self, ip, login, password, os, path_to_ftp_client, benchmark_config, log_file, res_file):
-        self.__computers.append(RemoteComputer(ip, login, password, os, path_to_ftp_client, benchmark_config,
-                                               log_file, res_file))
-
-    def change_computer(self, row, ip, login, password, os, path_to_ftp_client, benchmark_config, log_file, res_file):
-        self.__computers[row] = RemoteComputer(ip, login, password, os, path_to_ftp_client, benchmark_config,
-                                               log_file, res_file)
-
-    def delete_computer(self, index):
-        self.__computers.pop(index)
-
-    def delete_computers(self, indexes):
-        for index in indexes:
-            if index < len(self.__computers):
-                self.delete_computer(index)
-
-    def clear(self):
-        self.__computers.clear()
-
-    def parse_config(self, path_to_config):
-        CONFIG_ROOT_TAG = 'Computer'
-        CONFIG_IP_TAG = 'IP'
-        CONFIG_LOGIN_TAG = 'Login'
-        CONFIG_PASSWORD_TAG = 'Password'
-        CONFIG_OS_TAG = 'OS'
-        CONFIG_FTP_CLIENT_PATH_TAG = 'FTPClientPath'
-        CONFIG_BENCHMARK_CONFIG_TAG = 'BenchmarkConfig'
-        CONFIG_LOG_FILE_TAG = 'LogFile'
-        CONFIG_RESULT_FILE_TAG = 'ResultFile'
-
-        parsed_config = minidom.parse(path_to_config)
-        computers = parsed_config.getElementsByTagName(CONFIG_ROOT_TAG)
-        self.__computers.clear()
-        for idx, computer in enumerate(computers):
-            self.__computers.append(RemoteComputer())
-            self.__computers[idx].ip = computer.getElementsByTagName(CONFIG_IP_TAG)[0].firstChild.data
-            self.__computers[idx].login = computer.getElementsByTagName(CONFIG_LOGIN_TAG)[0].firstChild.data
-            self.__computers[idx].password = computer.getElementsByTagName(CONFIG_PASSWORD_TAG)[0].firstChild.data
-            self.__computers[idx].os = computer.getElementsByTagName(CONFIG_OS_TAG)[0].firstChild.data
-            self.__computers[idx].path_to_ftp_client = computer.getElementsByTagName(CONFIG_FTP_CLIENT_PATH_TAG)[0].firstChild.data
-            self.__computers[idx].benchmark_config = computer.getElementsByTagName(CONFIG_BENCHMARK_CONFIG_TAG)[0].firstChild.data
-            self.__computers[idx].log_file = computer.getElementsByTagName(CONFIG_LOG_FILE_TAG)[0].firstChild.data
-            self.__computers[idx].res_file = computer.getElementsByTagName(CONFIG_RESULT_FILE_TAG)[0].firstChild.data
-
-    def create_config(self, path_to_config):
-        if len(self.__computers) == 0:
-            return False
-        file = minidom.Document()
-        CONFIG_ROOT_TAG = file.createElement('Computers')
-        file.appendChild(CONFIG_ROOT_TAG)
-        for computer in self.__computers:
-            CONFIG_COMPUTER_TAG = file.createElement('Computer')
-            CONFIG_IP_TAG = file.createElement('IP')
-            CONFIG_LOGIN_TAG = file.createElement('Login')
-            CONFIG_PASSWORD_TAG = file.createElement('Password')
-            CONFIG_OS_TAG = file.createElement('OS')
-            CONFIG_FTP_CLIENT_PATH_TAG = file.createElement('FTPClientPath')
-            CONFIG_BENCHMARK_CONFIG_TAG = file.createElement('BenchmarkConfig')
-            CONFIG_LOG_FILE_TAG = file.createElement('LogFile')
-            CONFIG_RESULT_FILE_TAG = file.createElement('ResultFile')
-
-            CONFIG_IP_TAG.appendChild(file.createTextNode(computer.ip))
-            CONFIG_LOGIN_TAG.appendChild(file.createTextNode(computer.login))
-            CONFIG_PASSWORD_TAG.appendChild(file.createTextNode(computer.password))
-            CONFIG_OS_TAG.appendChild(file.createTextNode(computer.os))
-            CONFIG_FTP_CLIENT_PATH_TAG.appendChild(file.createTextNode(computer.path_to_ftp_client))
-            CONFIG_BENCHMARK_CONFIG_TAG.appendChild(file.createTextNode(computer.benchmark_config))
-            CONFIG_LOG_FILE_TAG.appendChild(file.createTextNode(computer.log_file))
-            CONFIG_RESULT_FILE_TAG.appendChild(file.createTextNode(computer.res_file))
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_IP_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_LOGIN_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_PASSWORD_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_OS_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_FTP_CLIENT_PATH_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_BENCHMARK_CONFIG_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_LOG_FILE_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_RESULT_FILE_TAG)
-            CONFIG_ROOT_TAG.appendChild(CONFIG_COMPUTER_TAG)
-        xml_str = file.toprettyxml(indent="\t", encoding="utf-8")
-        with open(path_to_config, 'wb') as f:
-            f.write(xml_str)
-        try:
-            f = open(path_to_config, 'r')
-            f.close()
-        except IOError:
-            return False
-        return True
-
-
-class DeployConfig:
-    def __init__(self):
-        self.__computers = []
-
-    def get_computers(self):
-        return self.__computers
-
-    def add_computer(self, ip, login, password, os, download_folder):
-        self.__computers.append(DeployComputer(ip, login, password, os, download_folder))
-
-    def change_computer(self, row, ip, login, password, os, download_folder):
-        self.__computers[row] = DeployComputer(ip, login, password, os, download_folder)
-
-    def delete_computer(self, index):
-        self.__computers.pop(index)
-
-    def delete_computers(self, indexes):
-        for index in indexes:
-            if index < len(self.__computers):
-                self.delete_computer(index)
-
-    def clear(self):
-        self.__computers.clear()
-
-    def parse_config(self, path_to_config):
-        CONFIG_ROOT_TAG = 'Computer'
-        CONFIG_IP_TAG = 'IP'
-        CONFIG_LOGIN_TAG = 'Login'
-        CONFIG_PASSWORD_TAG = 'Password'
-        CONFIG_OS_TAG = 'OS'
-        CONFIG_DOWNLOAD_FOLDER_TAG = 'DownloadFolder'
-
-        parsed_config = minidom.parse(path_to_config)
-        computers = parsed_config.getElementsByTagName(CONFIG_ROOT_TAG)
-        self.__computers.clear()
-        for idx, computer in enumerate(computers):
-            self.__computers.append(DeployComputer())
-            self.__computers[idx].ip = computer.getElementsByTagName(CONFIG_IP_TAG)[0].firstChild.data
-            self.__computers[idx].login = computer.getElementsByTagName(CONFIG_LOGIN_TAG)[0].firstChild.data
-            self.__computers[idx].password = computer.getElementsByTagName(CONFIG_PASSWORD_TAG)[0].firstChild.data
-            self.__computers[idx].os = computer.getElementsByTagName(CONFIG_OS_TAG)[0].firstChild.data
-            self.__computers[idx].download_folder = computer.getElementsByTagName(CONFIG_DOWNLOAD_FOLDER_TAG)[0].firstChild.data
-
-    def create_config(self, path_to_config):
-        if len(self.__computers) == 0:
-            return False
-        file = minidom.Document()
-        CONFIG_ROOT_TAG = file.createElement('Computers')
-        file.appendChild(CONFIG_ROOT_TAG)
-        for computer in self.__computers:
-            CONFIG_COMPUTER_TAG = file.createElement('Computer')
-            CONFIG_IP_TAG = file.createElement('IP')
-            CONFIG_LOGIN_TAG = file.createElement('Login')
-            CONFIG_PASSWORD_TAG = file.createElement('Password')
-            CONFIG_OS_TAG = file.createElement('OS')
-            CONFIG_DOWNLOAD_FOLDER_TAG = file.createElement('DownloadFolder')
-
-            CONFIG_IP_TAG.appendChild(file.createTextNode(computer.ip))
-            CONFIG_LOGIN_TAG.appendChild(file.createTextNode(computer.login))
-            CONFIG_PASSWORD_TAG.appendChild(file.createTextNode(computer.password))
-            CONFIG_OS_TAG.appendChild(file.createTextNode(computer.os))
-            CONFIG_DOWNLOAD_FOLDER_TAG.appendChild(file.createTextNode(computer.download_folder))
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_IP_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_LOGIN_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_PASSWORD_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_OS_TAG)
-            CONFIG_COMPUTER_TAG.appendChild(CONFIG_DOWNLOAD_FOLDER_TAG)
-            CONFIG_ROOT_TAG.appendChild(CONFIG_COMPUTER_TAG)
-        xml_str = file.toprettyxml(indent="\t", encoding="utf-8")
-        with open(path_to_config, 'wb') as f:
-            f.write(xml_str)
-        try:
-            f = open(path_to_config, 'r')
-            f.close()
-        except IOError:
-            return False
-        return True
-
-
-class DataBase:
-    def __init__(self):
-        self.models = Models()
-        self.data = Data()
-        self.benchmark_config = BenchmarkConfig()
-        self.remote_config = RemoteConfig()
-        self.deploy_config = DeployConfig()
