+++ gui/model/database.py
@@ -1,0 +1,14 @@
+from model.models.models import Models
+from model.data.data import Data
+from model.benchmark_config.benchmark_config import BenchmarkConfig
+from model.remote_config.remote_config import RemoteConfig
+from model.deploy_config.deploy_config import DeployConfig
+
+
+class DataBase:
+    def __init__(self):
+        self.models = Models()
+        self.data = Data()
+        self.benchmark_config = BenchmarkConfig()
+        self.remote_config = RemoteConfig()
+        self.deploy_config = DeployConfig()
