+++ gui/model/deploy_config/deploy_computer.py
@@ -1,0 +1,7 @@
+class DeployComputer:
+    def __init__(self, ip=None, login=None, password=None, os=None, download_folder=None):
+        self.ip = ip
+        self.login = login
+        self.password = password
+        self.os = os
+        self.download_folder = download_folder
