+++ gui/view/buttons/group_buttons.py
@@ -1,5 +1,4 @@
-from PyQt5.QtWidgets import *
-from PyQt5 import QtCore
+from PyQt5.QtWidgets import QGroupBox, QPushButton, QVBoxLayout
 
 
 class GroupButtons(QGroupBox):
@@ -22,10 +21,3 @@ class GroupButtons(QGroupBox):
 
     def get_buttons(self):
         return self.__buttons
-
-
-class DataGroupButtons(GroupButtons):
-    def _create_dict_buttons(self):
-        list_name_buttons = ['Add information', 'Delete information', 'Change information', 'Clear table']
-        buttons = {list_name_buttons[i]: QPushButton(list_name_buttons[i]) for i in range(len(list_name_buttons))}
-        return buttons
