+++ gui/view/dialogs/benchmark_config_dialog.py
@@ -1,117 +1,7 @@
-from PyQt5.QtWidgets import *
+from PyQt5.QtWidgets import QDialog, QLabel, QComboBox, QLineEdit, QPushButton, QGridLayout, QMessageBox
 
 
-class ModelDialog(QDialog):
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.__title = 'Information about model'
-        self.tags = ['Task', 'Name', 'Presicion', 'SourceFramework', 'ModelPath', 'WeightsPath']
-        self.__init_ui()
-
-    def __init_ui(self):
-        self.setWindowTitle(self.__title)
-        self.__create_labels()
-        self.__create_edits()
-        self.__create_layout()
-
-    def __create_labels(self):
-        self.labels = dict.fromkeys(self.tags)
-        for key in self.labels:
-            self.labels[key] = QLabel(key)
-
-    def __create_edits(self):
-        self.edits = dict.fromkeys(self.tags)
-        for key in self.edits:
-            self.edits[key] = QLineEdit(self)
-
-    def __create_layout(self):
-        layout = QGridLayout()
-        idx = 0
-        for tag in self.tags:
-            layout.addWidget(self.labels[tag], idx, 0)
-            layout.addWidget(self.edits[tag], idx, 1)
-            idx += 1
-        ok_btn = QPushButton('Ok')
-        cancel_btn = QPushButton('Cancel')
-        ok_btn.clicked.connect(self.accept)
-        cancel_btn.clicked.connect(self.reject)
-        layout.addWidget(ok_btn, idx, 0)
-        layout.addWidget(cancel_btn, idx, 1)
-        self.setLayout(layout)
-
-    def get_values(self):
-        values = []
-        for tag in self.tags:
-            values.append(self.edits[tag].text())
-        return values
-
-    def accept(self):
-        check = False
-        for tag in self.tags:
-            if self.edits[tag].text() == '':
-                check = True
-        if check:
-            QMessageBox.warning(self, 'Warning!', 'Not all lines are filled!')
-        else:
-            super().accept()
-
-
-class DataDialog(QDialog):
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.__title = 'Information about dataset'
-        self.tags = ['Name', 'Path']
-        self.__init_ui()
-
-    def __init_ui(self):
-        self.setWindowTitle(self.__title)
-        self.__create_labels()
-        self.__create_edits()
-        self.__create_layout()
-
-    def __create_labels(self):
-        self.labels = dict.fromkeys(self.tags)
-        for key in self.labels:
-            self.labels[key] = QLabel(key)
-
-    def __create_edits(self):
-        self.edits = dict.fromkeys(self.tags)
-        for key in self.edits:
-            self.edits[key] = QLineEdit(self)
-
-    def __create_layout(self):
-        layout = QGridLayout()
-        idx = 0
-        for tag in self.tags:
-            layout.addWidget(self.labels[tag], idx, 0)
-            layout.addWidget(self.edits[tag], idx, 1)
-            idx += 1
-        ok_btn = QPushButton('Ok')
-        cancel_btn = QPushButton('Cancel')
-        ok_btn.clicked.connect(self.accept)
-        cancel_btn.clicked.connect(self.reject)
-        layout.addWidget(ok_btn, idx, 0)
-        layout.addWidget(cancel_btn, idx, 1)
-        self.setLayout(layout)
-
-    def get_values(self):
-        values = []
-        for tag in self.tags:
-            values.append(self.edits[tag].text())
-        return values
-
-    def accept(self):
-        check = False
-        for tag in self.tags:
-            if self.edits[tag].text() == '':
-                check = True
-        if check:
-            QMessageBox.warning(self, 'Warning!', 'Not all lines are filled!')
-        else:
-            super().accept()
-
-
-class BenchmarkDialog(QDialog):
+class BenchmarkConfigDialog(QDialog):
     def __init__(self, parent, models, data):
         super().__init__(parent)
         self.framework_independent_tags = ['FrameworkIndependent:', 'Model', 'Dataset', 'Framework', 'BatchSize',
@@ -268,114 +158,6 @@ class BenchmarkDialog(QDialog):
         check = False
         for tag in self.framework_independent_tags[4:]:
             if tag != 'Device' and self.framework_independent_edits[tag].text() == '':
-                check = True
-        if check:
-            QMessageBox.warning(self, 'Warning!', 'Not all lines are filled!')
-        else:
-            super().accept()
-
-
-class RemoteDialog(QDialog):
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.tags = ['IP', 'Login', 'Password', 'OS', 'FTPClientPath', 'BenchmarkConfig', 'LogFile', 'ResultFile']
-        self.__init_ui()
-
-    def __init_ui(self):
-        self.setWindowTitle('Information about computer')
-        self.__create_labels()
-        self.__create_edits()
-        self.__create_layout()
-
-    def __create_labels(self):
-        self.labels = dict.fromkeys(self.tags)
-        for key in self.labels:
-            self.labels[key] = QLabel(key)
-
-    def __create_edits(self):
-        self.edits = dict.fromkeys(self.tags)
-        for key in self.edits:
-            self.edits[key] = QLineEdit(self)
-
-    def __create_layout(self):
-        layout = QGridLayout()
-        idx = 0
-        for tag in self.tags:
-            layout.addWidget(self.labels[tag], idx, 0)
-            layout.addWidget(self.edits[tag], idx, 1)
-            idx += 1
-        ok_btn = QPushButton('Ok')
-        cancel_btn = QPushButton('Cancel')
-        ok_btn.clicked.connect(self.accept)
-        cancel_btn.clicked.connect(self.reject)
-        layout.addWidget(ok_btn, idx, 0)
-        layout.addWidget(cancel_btn, idx, 1)
-        self.setLayout(layout)
-
-    def get_values(self):
-        values = []
-        for tag in self.tags:
-            values.append(self.edits[tag].text())
-        return values
-
-    def accept(self):
-        check = False
-        for tag in self.tags:
-            if self.edits[tag].text() == '':
-                check = True
-        if check:
-            QMessageBox.warning(self, 'Warning!', 'Not all lines are filled!')
-        else:
-            super().accept()
-
-
-class DeployDialog(QDialog):
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.tags = ['IP', 'Login', 'Password', 'OS', 'Download folder']
-        self.__init_ui()
-
-    def __init_ui(self):
-        self.setWindowTitle('Information about computer')
-        self.__create_labels()
-        self.__create_edits()
-        self.__create_layout()
-
-    def __create_labels(self):
-        self.labels = dict.fromkeys(self.tags)
-        for key in self.labels:
-            self.labels[key] = QLabel(key)
-
-    def __create_edits(self):
-        self.edits = dict.fromkeys(self.tags)
-        for key in self.edits:
-            self.edits[key] = QLineEdit(self)
-
-    def __create_layout(self):
-        layout = QGridLayout()
-        idx = 0
-        for tag in self.tags:
-            layout.addWidget(self.labels[tag], idx, 0)
-            layout.addWidget(self.edits[tag], idx, 1)
-            idx += 1
-        ok_btn = QPushButton('Ok')
-        cancel_btn = QPushButton('Cancel')
-        ok_btn.clicked.connect(self.accept)
-        cancel_btn.clicked.connect(self.reject)
-        layout.addWidget(ok_btn, idx, 0)
-        layout.addWidget(cancel_btn, idx, 1)
-        self.setLayout(layout)
-
-    def get_values(self):
-        values = []
-        for tag in self.tags:
-            values.append(self.edits[tag].text())
-        return values
-
-    def accept(self):
-        check = False
-        for tag in self.tags:
-            if self.edits[tag].text() == '':
                 check = True
         if check:
             QMessageBox.warning(self, 'Warning!', 'Not all lines are filled!')
