+++ gui/view/view.py
@@ -1,0 +1,15 @@
+from PyQt5.QtWidgets import QMainWindow
+from view.widgets.main_widget import MainWidget
+
+
+class View(QMainWindow):
+    def __init__(self):
+        super().__init__()
+        self.resize(920, 560)
+        self.setWindowTitle('Configuration Creator')
+        self.tabs = MainWidget(self)
+        self.setCentralWidget(self.tabs)
+        self.show()
+
+    def update(self, model):
+        self.tabs.update(model)
